+++ src/interfaces/ILogger.ts
@@ -1,0 +1,6 @@
+export interface ILogger {
+    debug(...arg: unknown[]): void;
+    info(...arg: unknown[]): void;
+    warn(...arg: unknown[]): void;
+    error(...arg: unknown[]): void;
+}