+++ tests/unit/sdk.spec.ts
@@ -8,7 +8,7 @@ import { mnemonicNew } from 'ton-crypto';
 
 import { Asset, EvmProxyMsg, Network, SenderFactory, TacSdk, wallets, WalletVersion } from '../../src';
 import { sandboxOpener } from '../../src/adapters/contractOpener';
-import { TON } from '../../src/assets';
+import { TON } from '../../src';
 
 describe('TacSDK', () => {
     const {